+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to `livewire-filters` will be documented in this file.
 
+## 0.3 - 2022-03-05
+
+- Added `getOptionId` method for better handling of setting an option's ID attribute when dealing with associative arrays
+- Added `getOptionName` method for better handling of setting an option's name attribute when dealing with associative arrays
+- Added `getOptionValue` method for better handling of setting an option's value when dealing with associative arrays
+
 ## 0.2 - 2022-02-28
 
 - Added the `meta` field to `Filter` class